+++ smoke-tests/apps/SpringScheduling/src/main/java/com/microsoft/applicationinsights/smoketestapp/TestController.java
@@ -3,11 +3,16 @@
 
 package com.microsoft.applicationinsights.smoketestapp;
 
+import java.time.Instant;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.scheduling.TaskScheduler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class TestController {
+
+  @Autowired private TaskScheduler taskScheduler;
 
   @GetMapping("/")
   public String root() {
@@ -16,6 +21,12 @@ public class TestController {
 
   @GetMapping("/scheduler")
   public String scheduler() {
+    return "OK!";
+  }
+
+  @GetMapping("/should-ignore")
+  public String shouldIgnoreTest() throws Exception {
+    taskScheduler.schedule(() -> System.out.println("here i am"), Instant.now());
     return "OK!";
   }
 }